test: factor filesystem fixtures into tests/fixtures, wire via pytest_plugins
Move confine_is_file_to out of conftest.py into tests/fixtures/filesystem.py,
registered project-wide through pytest_plugins. Keeps conftest focused on node
registration and execution-algorithm fixtures, and gives filesystem/CLI test
helpers a dedicated home as they accumulate.
