+++ tests/conftest.py
@@ -6,8 +6,7 @@ node registry, so anything declared here is available to every test file
 without needing to be re-imported.
 """
 
-from pathlib import Path
-from typing import Callable, ClassVar, Type
+from typing import ClassVar, Type
 
 import pytest
 from overrides import override
@@ -25,6 +24,12 @@ from workflow_engine import (
 )
 from workflow_engine.execution import TopologicalExecutionAlgorithm
 from workflow_engine.execution.parallel import ParallelExecutionAlgorithm
+
+# Fixture modules registered project-wide. Add new fixture files here rather
+# than letting this conftest sprawl.
+pytest_plugins = [
+    "tests.fixtures.filesystem",
+]
 
 
 @pytest.fixture(params=["topological", "parallel"])
@@ -79,33 +84,6 @@ class YieldingNode(Node[Empty, SimpleOutput, Params]):
     ) -> SimpleOutput:
         YieldingNode.calls[self.id] = YieldingNode.calls.get(self.id, 0) + 1
         raise ShouldYield(self.message)
-
-
-@pytest.fixture
-def confine_is_file_to(monkeypatch) -> Callable[[Path], None]:
-    """Confine `Path.is_file()` to a subtree, hiding everything outside it.
-
-    CLI helpers that walk up the filesystem looking for marker files
-    (`engine.yaml`, `pyproject.toml`, ...) ascend all the way to the root, so
-    a stray file in a real parent directory (e.g. `/tmp` or a CI home dir) can
-    make "not found" assertions flaky. Call the returned function with a root
-    (typically `tmp_path`): paths at or below it delegate to the real
-    `is_file`, everything else reports `False`.
-    """
-    real_is_file = Path.is_file
-
-    def confine(root: Path) -> None:
-        resolved_root = root.resolve()
-
-        def fake(self: Path) -> bool:
-            resolved = self.resolve()
-            if resolved != resolved_root and resolved_root not in resolved.parents:
-                return False
-            return real_is_file(self)
-
-        monkeypatch.setattr(Path, "is_file", fake)
-
-    return confine
 
 
 @pytest.fixture(autouse=True)
